Remove user from source_users and test it

+++ src/pas/plugins/imio/plugin.py
@@ -3,7 +3,6 @@ from AccessControl import ClassSecurityInfo
 from App.class_init import InitializeClass
 from operator import itemgetter
 from pas.plugins.authomatic.plugin import AuthomaticPlugin
-from pas.plugins.authomatic.useridfactories import new_userid
 from pas.plugins.imio.interfaces import IAuthenticPlugin
 from plone import api
 from Products.PageTemplates.PageTemplateFile import PageTemplateFile
@@ -59,10 +58,13 @@ class AuthenticPlugin(AuthomaticPlugin):
         useridentities = super(AuthenticPlugin, self).remember_identity(result, userid)
         if userid is None:
             # remove old plone userid from source_users
-            userid = new_userid(self, result)
-            if self._useridentities_by_userid.get(userid, None):
-                acl_users = api.portal.get_tool("acl_users")
-                source_users = acl_users.source_users
+            userid = result.user.username
+            acl_users = api.portal.get_tool("acl_users")
+            source_users = acl_users.source_users
+            if (
+                self._useridentities_by_userid.get(userid, None)
+                and len(source_users.enumerateUsers(userid)) > 0
+            ):
                 source_users.doDeleteUser(userid)
         return useridentities
